Various last minute improvements (#21)
* delete ledge, as was never implemented

* update interface docs to reflect stateful game

* small docstring changes to nim & hex

* move LiteModel dummy prediction into init of nn

* visualization for topp games (as per project description)

Quoting the project description:

During any phase of system operation, whether training or TOPP play, it must be possible to view the progress of
individual games via this graphic. This does not include individual rollout simulations, but does include each actual
game, i.e. episode, during training and each game of a tournament.
This display feature must be easy to turn on/off by the user. During the demonstration session, we will want to watch
a few actual games in progress....but not all games.